new - corrigindo testes que quebraram

+++ recipes/signals.py
@@ -7,13 +7,16 @@ from utils.delete_media_files import delete_file
 
 @receiver(pre_delete, sender=Recipe)
 def recipe_cover_delete(sender, instance, *args, **kwargs):
-    old_instance = Recipe.objects.filter(pk=instance.pk)
+    old_instance = Recipe.objects.filter(pk=instance.pk).first()
     delete_file(old_instance)
 
 
 @receiver(pre_save, sender=Recipe)
 def recipe_cover_update(sender, instance, *args, **kwargs):
-    old_instance = Recipe.objects.filter(pk=instance.pk)
+    old_instance = Recipe.objects.filter(pk=instance.pk).first()
+
+    if not old_instance:
+        return
 
     if old_instance.cover != instance.cover:
         delete_file(old_instance)
